requirements.txt
PyYAML
pybit
elasticsearch
